requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.39.0
google-generativeai==0.8.3
openai==1.51.0
supabase==2.9.1
pydantic==2.9.2
httpx==0.27.2
